(WIP) Move currency conversions to own module.
