+++ routines.py
@@ -5,36 +5,19 @@ import sys
 import requests
 
 import paket_stellar
-import util.conversion
 import util.logger
 
+import currency_conversions
 import db
 
 LOGGER = util.logger.logging.getLogger('pkt.funder.routines')
 DEBUG = bool(os.environ.get('PAKET_DEBUG'))
 FUNDER_SEED = os.environ['PAKET_FUNDER_SEED']
-BUL_STROOPS_FOR_EUR_CENT = int(os.environ['PAKET_BUL_STROOPS_FOR_EUR_CENT'])
 ETHERSCAN_API_KEY = os.environ['PAKET_ETHERSCAN_API_KEY']
-# currencies ids on coinmarketcap.com
-XLM_ID = 512
-ETH_ID = 1027
-BTC_ID = 1
 
 
 class BalanceError(Exception):
     """Can't get balance for specified address"""
-
-
-def get_currency_price(id_, convert):
-    """
-    Get crypto currency price in specified fiat currency.
-    Crypto currency specifies as id from coinmarketcap.com
-    """
-    url = 'https://api.coinmarketcap.com/v2/ticker/{}/?convert={}'.format(id_, convert)
-    response = requests.get(url)
-    price = response.json()['data']['quotes'][convert]['price']
-    # we need to cast to string because API returns price as float number
-    return str(price)
 
 
 def get_btc_balance(address):
@@ -68,35 +51,6 @@ def get_balance(address, network):
     return get_btc_balance(address) if network == 'BTC' else get_eth_balance(address)
 
 
-def currency_to_euro_cents(currency, amount):
-    """Convert amount of coins in specified currency to euro cents"""
-    assert currency in ['BTC', 'ETH'], 'currency must be BTC or ETH'
-    id_, decimals = (BTC_ID, util.conversion.BTC_DECIMALS) if currency == 'BTC' else (
-        ETH_ID, util.conversion.ETH_DECIMALS)
-    eur_price = get_currency_price(id_, 'EUR')
-    price_decimals = len(eur_price.split('.')[1])
-    # price in fictitious units (portions of euro cents) by 1 indivisible unit of specified crypto currency
-    fictitious_units_price = util.conversion.divisible_to_indivisible(eur_price, price_decimals)
-    fictitious_units_amount = fictitious_units_price * amount
-    # minus two because initial price was in EUR and we want euro cents
-    euro_cents = util.conversion.indivisible_to_divisible(fictitious_units_amount, price_decimals + decimals - 2)
-    LOGGER.warning("precision loss: %s converted to %s", euro_cents, round(float(euro_cents)))
-    # integer part of result will be amount of euro cents
-    return round(float(euro_cents))
-
-
-def euro_cents_to_stroops(euro_cents_amount):
-    """Convert amount of euro cents to stroops"""
-    eur_price = get_currency_price(XLM_ID, 'EUR')
-    price_decimals = len(eur_price.split('.')[1])
-    fictitious_units_amount = util.conversion.divisible_to_indivisible(
-        euro_cents_amount, util.conversion.STELLAR_DECIMALS + price_decimals)
-    fictitious_units_price = util.conversion.divisible_to_indivisible(eur_price, price_decimals + 2)
-    stroops = fictitious_units_amount // fictitious_units_price
-    LOGGER.warning("precision loss: %s / %s = %s", fictitious_units_amount, fictitious_units_price, stroops)
-    return stroops
-
-
 def fund_account(user_pubkey, amount, asset_code):
     """Fund account with XLM or BUL"""
     assert asset_code in ['XLM', 'BUL'], 'asset must be XLM or BUL'
@@ -117,7 +71,7 @@ def check_purchases_addresses():
     for purchase in purchases:
         LOGGER.info("checking address %s", purchase['payment_pubkey'])
         balance = get_balance(purchase['payment_pubkey'], purchase['payment_currency'])
-        euro_cents_balance = currency_to_euro_cents(purchase['payment_currency'], balance)
+        euro_cents_balance = currency_conversions.currency_to_euro_cents(purchase['payment_currency'], balance)
         if euro_cents_balance >= db.MINIMUM_PAYMENT:
             db.update_purchase(purchase['payment_pubkey'], 1)
 
@@ -127,14 +81,14 @@ def send_requested_currency():
     purchases = db.get_paid()
     for purchase in purchases:
         balance = get_balance(purchase['payment_pubkey'], purchase['payment_currency'])
-        euro_cents_balance = currency_to_euro_cents(purchase['payment_currency'], balance)
+        euro_cents_balance = currency_conversions.currency_to_euro_cents(purchase['payment_currency'], balance)
         monthly_allowance = db.get_monthly_allowance(purchase['user_pubkey'])
         monthly_expanses = db.get_monthly_expanses(purchase['user_pubkey'])
         remaining_monthly_allowance = monthly_allowance - monthly_expanses
         euro_to_fund = min(euro_cents_balance, remaining_monthly_allowance)
         if euro_to_fund:
             if purchase['requested_currency'] == 'BUL':
-                fund_amount = euro_to_fund * BUL_STROOPS_FOR_EUR_CENT
+                fund_amount = currency_conversions.euro_cents_to_bul_stroops(euro_to_fund)
                 try:
                     account = paket_stellar.get_bul_account(purchase['user_pubkey'])
                     if account['bul_balance'] + fund_amount <= account['bul_limit']:
@@ -150,7 +104,7 @@ def send_requested_currency():
                     LOGGER.error(str(exc))
                     db.update_purchase(purchase['payment_pubkey'], -1)
             else:
-                fund_amount = euro_cents_to_stroops(euro_to_fund)
+                fund_amount = currency_conversions.euro_cents_to_xlm_stroops(euro_to_fund)
                 try:
                     paket_stellar.get_bul_account(purchase['user_pubkey'], accept_untrusted=True)
                     fund_account(purchase['user_pubkey'], fund_amount, 'XLM')
